Feature/demo/demo movie/yahata

+++ Assets/Scripts/Shaders/SilToon/HLSL/Fragment/FaceLight.hlsl
@@ -1,7 +1,7 @@
 #ifndef FACE_LIGHT_INCLUDED
 #define FACE_LIGHT_INCLUDED
 
-float3 GetFaceNormal(float3 faceUpWS, float3 normalWS)
+half3 GetFaceNormal(half3 faceUpWS, half3 normalWS)
 {
     return normalize(-dot(faceUpWS, normalWS) * faceUpWS + normalWS);
 }
